backend/requirements.txt
openai
python-dotenv
psycopg2-binary
# python == 3.12.7 
fastapi==0.115.2
uvicorn==0.31.1
aiohttp==3.10.10    
requests==2.32.3    
python-dotenv==1.0.1
google-generativeai==0.8.3
SQLAlchemy==2.0.36
PyJWT==2.9.0
jsonschema==4.23
referencing==0.35
sqlglot